bug: remove config and dotenv flow dependency from useChain hook
