+++ packages/sdk/src/config/createContext.ts
@@ -8,35 +8,10 @@ import {
 } from "viem";
 import { privateKeyToAccount } from "viem/accounts";
 import { getChainConfirmations } from "./utils/getChainConfirmations";
-import * as chains from "viem/chains";
 import type { VlayerContextConfig } from "./types";
+import { getChainSpecs } from "../utils";
 
 export type EthClient = ReturnType<typeof createContext>["ethClient"];
-
-const isChain = (chain: unknown): chain is Chain => {
-  return (
-    typeof chain === "object" &&
-    chain !== null &&
-    "id" in chain &&
-    "name" in chain &&
-    "nativeCurrency" in chain &&
-    "rpcUrls" in chain
-  );
-};
-
-export const getChainSpecs = (chainName: string): Chain => {
-  let chain = undefined;
-  try {
-    chain = chains[chainName as keyof typeof chains];
-  } catch {
-    throw Error(`Cannot import ${chainName} from viem/chains`);
-  }
-
-  if (!chain || !isChain(chain)) {
-    throw new Error(`Chain ${chainName} is not supported by viem`);
-  }
-  return chain;
-};
 
 export const customTransport = custom;
