+++ packages/sdk/src/utils/index.ts
@@ -1,0 +1,1 @@
+export { getChainSpecs } from "./getChainSpecs";
